ai/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.10.0
pydantic-settings==2.5.2

# LLM
openai==1.52.0
google-generativeai==0.8.3

# RAG & Embeddings
sentence-transformers==3.2.1
faiss-cpu==1.9.0
numpy==2.1.3

# Data Processing
pandas==2.2.3
openpyxl==3.1.5

# Security & Validation
python-jose[cryptography]==3.3.0
httpx==0.27.2
requests==2.32.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
tenacity==9.0.0
structlog==24.4.0
